+++ changelog.md
@@ -6,6 +6,7 @@
   - [PRE-2579](https://gitlab.com/dalenys/public/ecommerce/prestashop_v4_17_8/-/merge_requests/2526): Add queue system on notifications and flag it
   - [PRE-2581](https://gitlab.com/dalenys/public/ecommerce/prestashop_v4_17_8/-/merge_requests/2531): Add queue system on payment capture and flag it
   - [PRE-2580](https://gitlab.com/dalenys/public/ecommerce/prestashop_v4_17_8/-/merge_requests/2532): Add queue system on refund and flag it
+  - [PRE-2578](https://gitlab.com/dalenys/public/ecommerce/prestashop_v4_17_8/-/merge_requests/2527): Add queue system on validation and flag it
 
 ## Version 4.11.2
 - Bugfix :